infrastructure/cdk/requirements.txt
# AWS CDK core
aws-cdk-lib==2.173.2
constructs>=10.0.0,<11.0.0

# CDK Nag for compliance checks
cdk-nag==2.34.0
